Added ScannerHost interface
Scanners should get their Options from this interface instead of passing around instances of options everywhere
