+++ Source/ZXing.Net.Mobile.Core/IScannerSessionHost.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace ZXing.Mobile
+{
+	public interface IScannerSessionHost
+	{
+		ZXing.Mobile.MobileBarcodeScanningOptions ScanningOptions { get; }
+	}
+}
